requirements.txt
streamlit
tensorflow
numpy
joblib
Pillow